tech?(api): move getnextchallengeforcampaignassessment usecase in evaluation bounded context
