+++ api/src/evaluation/domain/usecases/get-next-challenge-for-campaign-assessment.js
@@ -1,17 +1,17 @@
-import { FlashAssessmentAlgorithm } from '../../../src/certification/flash-certification/domain/models/FlashAssessmentAlgorithm.js';
-import { FlashAssessmentAlgorithmConfiguration } from '../../../src/certification/shared/domain/models/FlashAssessmentAlgorithmConfiguration.js';
-import { AssessmentEndedError } from '../../../src/shared/domain/errors.js';
+import { FlashAssessmentAlgorithm } from '../../../certification/flash-certification/domain/models/FlashAssessmentAlgorithm.js';
+import { FlashAssessmentAlgorithmConfiguration } from '../../../certification/shared/domain/models/FlashAssessmentAlgorithmConfiguration.js';
+import { AssessmentEndedError } from '../../../shared/domain/errors.js';
 
 const getNextChallengeForCampaignAssessment = async function ({
+  assessment,
+  locale,
   challengeRepository,
   answerRepository,
   flashAlgorithmConfigurationRepository,
   flashAssessmentResultRepository,
-  assessment,
   pickChallengeService,
-  locale,
   algorithmDataFetcherService,
-  smartRandom,
+  smartRandomService,
   flashAlgorithmService,
 }) {
   let algoResult;
@@ -45,7 +45,7 @@ const getNextChallengeForCampaignAssessment = async function ({
     return pickChallengeService.chooseNextChallenge(assessment.id)({ possibleChallenges });
   } else {
     const inputValues = await algorithmDataFetcherService.fetchForCampaigns(...arguments);
-    algoResult = smartRandom.getPossibleSkillsForNextChallenge({ ...inputValues, locale });
+    algoResult = smartRandomService.getPossibleSkillsForNextChallenge({ ...inputValues, locale });
 
     if (algoResult.hasAssessmentEnded) {
       throw new AssessmentEndedError();
